requirements.txt
boto3==1.9.71
botocore==1.12.71
PyYAML==3.13
MarkupSafe==1.1.0
jinja2==2.10